Add files via upload
This is a project from my Spark class where in we had to analyze a directory of 50,000+ food items sold in Belgium. One of the goals was to determine how much of the food in the directory could be considered 'healthy'. The coding was done primarily by myself and Eleonora Jimenez